Update API functions and introduce new test applications
- Modified the read (rd) and write (wr) functions in api.c to accept uint32_t buffers, enhancing data handling capabilities.
- Updated corresponding declarations in api.h to reflect the new buffer type.
- Added two new test applications, tiny_test.c and small_test.c, to validate memory operations and data integrity.
- Adjusted the Makefile to include the new test applications and ensure proper compilation.
- Enhanced utility functions in utils.c and utils.h to support the new buffer type and data pattern generation.

+++ src/software/api.h
@@ -9,7 +9,7 @@ void cleanup_hardware();
 
 uint32_t pre(uint8_t bank_addr, uint8_t rank_addr, bool bank_all, uint32_t interval, bool strict);
 uint32_t act(uint8_t bank_addr, uint32_t row_addr, uint8_t rank_addr, uint32_t interval, bool strict);
-uint32_t rd(uint8_t *buffer, uint8_t bank_addr, uint16_t col_addr, uint32_t interval, bool strict);
-uint32_t wr(uint8_t *buffer, uint8_t bank_addr, uint16_t col_addr, uint32_t interval, bool strict);
+uint32_t rd(uint32_t *buffer, uint8_t bank_addr, uint16_t col_addr, uint32_t interval, bool strict);
+uint32_t wr(uint32_t *buffer, uint8_t bank_addr, uint16_t col_addr, uint32_t interval, bool strict);
 
 #endif
